[FEATURE][2/2] Accessing ParameterBag values and using debug helper
Values configured in "parameters"-section of services.yaml are managed
via the ParameterBag. We can inject this ParameterBag to our e.g. Controller and access its values.

Also in this example, one value is being output in MyFirstController, using the "dump" PHP function. Values tracked with this function, can get viewed in the Debug profiler toolbar.