[TASK] Remove "products" from ProductCategoryType
We don't want to assign products to categories,
we want to choose the categories within the product.

To achieve this, we also need to implement __toString() method in
ProductCategory entity. The form component tries to provide a dropdown
list of available categories and for the labels to be shown, this is required.